Move I/O review code into ts/image-occlusion
A bit simpler when it's all in one place.
