+++ ts/reviewer/reviewer_extras.ts
@@ -5,11 +5,12 @@
 @typescript-eslint/no-explicit-any: "off",
  */
 
-// A standalone bundle that adds mutateNextCardStates to the anki namespace.
-// When all clients are using reviewer.js directly, we can get rid of this.
+// A standalone bundle that adds mutateNextCardStates and setupImageCloze
+// to the anki namespace. When all clients are using reviewer.js directly, we
+// can get rid of this.
 
+import { setupImageCloze } from "../image-occlusion/review";
 import { mutateNextCardStates } from "./answering";
-import { setupImageCloze } from "./image_occlusion";
 
 globalThis.anki = globalThis.anki || {};
 globalThis.anki.mutateNextCardStates = mutateNextCardStates;
